Created test_data folder with block json files for testing + create separate file for block base.
